[Refactor] refactoring for PCTSP
